Create control (espace de depot) : first version (#182)
