+++ ecc/urls.py
@@ -18,11 +18,12 @@ from user_profiles import api_views as user_profiles_api_views
 admin.site.site_header = 'e-contrôle Administration'
 
 router = routers.DefaultRouter()
+router.register(r'annexe', control_api_views.QuestionFileViewSet, basename='annexe')
+router.register(r'control', control_api_views.ControlViewSet, basename='control')
 router.register(r'question', control_api_views.QuestionViewSet, basename='question')
 router.register(r'questionnaire', control_api_views.QuestionnaireViewSet, basename='questionnaire')
 router.register(r'theme', control_api_views.ThemeViewSet, basename='theme')
 router.register(r'user', user_profiles_api_views.UserProfileViewSet, basename='user')
-router.register(r'annexe', control_api_views.QuestionFileViewSet, basename='annexe')
 
 urlpatterns = [
     path('', ecc_views.login, name='login'),
