+++ static/src/users/Users.vue
@@ -4,10 +4,10 @@
     <div class="card-header border-0" data-toggle="card-collapse">
       <div class="card-title">
         <i class="fe fe-users dropdown-icon"></i>
-        <a href="#">Personnes ayant accès au contrôle</a>
+        <a href="javascript:void(0)">Personnes ayant accès à l'espace</a>
       </div>
       <div class="card-options">
-        <a href="#" class="card-options-collapse"><i class="fe fe-chevron-up"></i></a>
+        <a href="javascript:void(0)" class="card-options-collapse"><i class="fe fe-chevron-up"></i></a>
       </div>
     </div>
 
@@ -18,7 +18,7 @@
         </div>
         <div class="col-auto">
           <a v-if="sessionUser.is_inspector"
-             href=""
+             href="javascript:void(0)"
              data-toggle="modal"
              data-target="#addUserModal"
              @click="updateEditingState('inspector')"
